Fix pip install --no-binary
It looks like local imports are not supported when installing with the
--no-binary flag, so fold in code from setuputils.py back into setup.py:

$ pip3 install --no-binary :all: --no-deps sqlalchemy-vertica-python
WARNING: Skipping sqlalchemy-vertica-python as it is not installed.
Looking in indexes: https://pypi.org/simple, https://reader:****@bluelabs.jfrog.io/bluelabs/api/pypi/pypi-local/simple
Collecting sqlalchemy-vertica-python
  Downloading sqlalchemy-vertica-python-0.5.6.tar.gz (5.1 kB)
    ERROR: Command errored out with exit status 1:
     command: /Users/broz/.pyenv/versions/3.7.4/envs/mylibs-3/bin/python -c 'import sys, setuptools, tokenize; sys.argv[0] = '"'"'/private/var/folders/x5/k18kxt7547n7tmkc0zkdyc7h0000gn/T/pip-install-j6zj_y_i/sqlalchemy-vertica-python/setup.py'"'"'; __file__='"'"'/private/var/folders/x5/k18kxt7547n7tmkc0zkdyc7h0000gn/T/pip-install-j6zj_y_i/sqlalchemy-vertica-python/setup.py'"'"';f=getattr(tokenize, '"'"'open'"'"', open)(__file__);code=f.read().replace('"'"'\r\n'"'"', '"'"'\n'"'"');f.close();exec(compile(code, __file__, '"'"'exec'"'"'))' egg_info --egg-base /private/var/folders/x5/k18kxt7547n7tmkc0zkdyc7h0000gn/T/pip-install-j6zj_y_i/sqlalchemy-vertica-python/pip-egg-info
         cwd: /private/var/folders/x5/k18kxt7547n7tmkc0zkdyc7h0000gn/T/pip-install-j6zj_y_i/sqlalchemy-vertica-python/
    Complete output (5 lines):
    Traceback (most recent call last):
      File "<string>", line 1, in <module>
      File "/private/var/folders/x5/k18kxt7547n7tmkc0zkdyc7h0000gn/T/pip-install-j6zj_y_i/sqlalchemy-vertica-python/setup.py", line 2, in <module>
        from setuputils import TestCoverageRatchetCommand, VerifyVersionCommand
    ModuleNotFoundError: No module named 'setuputils'
    ----------------------------------------
ERROR: Command errored out with exit status 1: python setup.py egg_info Check the logs for full command output.
$
