--- setuputils.py
@@ -1,72 +1,0 @@
-from decimal import Decimal
-from distutils.cmd import Command
-import os.path
-import sys
-
-from setuptools.command.install import install
-
-
-# From https://circleci.com/blog/continuously-deploying-python-packages-to-pypi-with-circleci/
-class VerifyVersionCommand(install):
-    """Custom command to verify that the git tag matches our version"""
-    description = 'verify that the git tag matches our version'
-
-    def run(self):
-        from setup import __version__
-        tag = os.getenv('CIRCLE_TAG')
-        tag_formatted_version = 'v{}'.format(__version__)
-
-        if tag != tag_formatted_version:
-            info = "Git tag: {0} does not match the version of this app: {1}".format(
-                tag, __version__
-            )
-            sys.exit(info)
-
-
-class CoverageRatchetCommand(Command):
-    description = 'Run coverage ratchet'
-    user_options = []  # type: ignore
-
-    def finalize_options(self) -> None:
-        pass
-
-    def run(self) -> None:
-        """Run command."""
-        import xml.etree.ElementTree as ET
-
-        tree = ET.parse(self.coverage_source_file)
-        new_coverage = Decimal(tree.getroot().attrib["line-rate"]) * 100
-
-        if not os.path.exists(self.coverage_file):
-            with open(self.coverage_file, 'w') as f:
-                f.write('0')
-
-        with open(self.coverage_file, 'r') as f:
-            high_water_mark = Decimal(f.read())
-
-        if new_coverage < high_water_mark:
-            raise Exception(
-                "{} coverage used to be {}; "
-                "down to {}%.  Fix by viewing '{}'".format(self.type_of_coverage,
-                                                           high_water_mark,
-                                                           new_coverage,
-                                                           self.coverage_url))
-        elif new_coverage > high_water_mark:
-            with open(self.coverage_file, 'w') as f:
-                f.write(str(new_coverage))
-            print("Just ratcheted coverage up to {}%".format(new_coverage))
-        else:
-            print("Code coverage steady at {}%".format(new_coverage))
-
-
-class TestCoverageRatchetCommand(CoverageRatchetCommand):
-    def initialize_options(self) -> None:
-        """Set default values for options."""
-        self.type_of_coverage = 'Test'
-        self.coverage_url = 'cover/index.html'
-        self.coverage_file = os.path.join(
-            os.path.dirname(os.path.abspath(__file__)),
-            'metrics',
-            'coverage_high_water_mark'
-        )
-        self.coverage_source_file = "coverage.xml"
